Create MinimumStops.cpp
Suppose you and your friends need to drive across the country. You want to drive as much as you can during the day, but decide that you will rest every night. You have identified a list of suitable sleeping points along the drive and need a system
for determining when to stop each evening. You decide to use a greedy approach.Design a greedy algorithm that minimizes the number of stops necessary for the trip